fix: split server fn impl into separate files to prevent client bundle errors
The _readSignalsHandler/_saveOverridesHandler/_readArchiveHandler functions
used fs/path at module level. Since they were exported from the server files,
TanStack Start's dead code elimination couldn't remove them from the client
bundle, causing "resolve is not exported by __vite-browser-external" errors.

Fix: extract pure impl logic into signals-impl.ts and archive-impl.ts. The
server files now only export createServerFn wrappers (with 'use server'
directive), and the tests import directly from the impl files.
